feat: updates due to testing
By adding some tests, there were some changes
that had to be made to the code itself.

`handleChange` can now directly take an `Event`
as an argument but, only if it has a `name` and a `value`
attribute. The old way, (using a fields object as the first,
and a validations array as the second argument) still works.

BREAKING CHANGE: form has to be
explicitly sent in to validate() function,
to be able to differentiate between new
and old values in validation.