fix(useForm): handle checked  in handleChange
The handleChange method can take a simple form event as an argument,
and the rest is handled internally.
However, in most cases, checkboxes take the
checked value as their value.
So if a an event comes from
a checkbox input, we set target.checked as a value,
as opposed to target.value.

Closes #4.