feat(FormProvider): submit, onNotify in FormProvider
It is now possible to pass onNotify and submit functions to FormProvider, instead of defining them in each useForm. The handleSubmit function now also gives the name of the form it was called from, to be able to identify what form called the submit function if it was passed to FormProvider.